server/requirements.txt
asgiref==3.3.4
backports.zoneinfo==0.2.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.1.0
MarkupSafe==1.1.1
numpy==1.22.1
packaging==20.9
pandas==1.3.5
Pillow==9.0.1
psycopg2==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
razorpay==1.3.0
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==3.0.0
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0